Mainrequirements.txt
tensorflow==1.15
umap-learn
visdom
librosa>=0.5.1
matplotlib>=2.0.2
numpy>=1.14.0
scipy>=1.0.0
tqdm
sounddevice
SoundFile
Unidecode
inflect
PyQt5
multiprocess
numba==0.48.0
torch
absl-py==0.8.1
annoy==1.16.2
appdirs==1.4.4
args==0.1.0
astor==0.8.0
atari-py==0.2.6
attrs==19.3.0
audioread==2.1.8
backcall==0.1.0
bleach==3.1.0
blis==0.4.1
box2d-py==2.3.8
cachetools==3.1.1
catalogue==0.0.8
certifi==2019.11.28
cffi==1.14.2
chardet==3.0.4
click==6.7
clint==0.5.1
cloudpickle==1.2.2
cudf-cuda100==0.6.1
cuml-cuda100==0.6.1
cupy==7.0.0
cupy-cuda100==7.0.0
cycler==0.10.0
cymem==2.0.3
Cython==0.29.14
cytoolz==0.10.1
decorator==4.4.1
defusedxml==0.6.0
dill==0.3.2
dlib==19.18.0
entrypoints==0.3
fastrlock==0.4
Flask==1.1.1
floyd-cli==0.11.17
funcsigs==1.0.2
future==0.18.2
gast==0.2.2
google-pasta==0.1.8
grpcio==1.25.0
gym==0.15.4
gym-retro==0.7.0
h5py==2.10.0
idna==2.8
imageio==2.6.1
importlib-metadata==1.2.0
incremental==17.5.0
inflect==4.1.0
ipykernel==5.1.3
ipython==7.10.1
ipython-genutils==0.2.0
ipywidgets==7.5.1
itsdangerous==1.1.0
jedi==0.15.1
jellyfish==0.7.2
Jinja2==2.10.3
joblib==0.14.0
json5==0.8.5
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==5.3.4
jupyter-console==6.0.0
jupyter-core==4.6.1
jupyterlab==1.2.3
jupyterlab-server==1.0.6
kaggle==1.5.6
Keras==2.3.1
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
kiwisolver==1.1.0
librosa==0.8.0
llvmlite==0.34.0
Markdown==3.1.1
MarkupSafe==1.1.1
marshmallow==2.20.5
matplotlib==3.1.2
menpo==0.9.2
mistune==0.8.4
mock==3.0.5
more-itertools==8.0.0
mpmath==1.1.0
multiprocess==0.70.10
murmurhash==1.0.2
nbconvert==5.6.1
nbformat==4.4.0
networkx==2.4
nltk==3.4.5
notebook==6.0.3
numba==0.51.2
numpy==1.17.4
nvstrings-cuda100==0.3.0.post1
opencv-contrib-python==3.4.0.12
opencv-python==4.1.2.30
opt-einsum==3.1.0
packaging==20.4
pandas==0.25.3
pandocfilters==1.4.2
parso==0.5.1
path==13.1.0
path.py==12.4.0
pathlib2==2.3.5
pbr==5.4.4
pexpect==4.7.0
pickleshare==0.7.5
Pillow==6.2.1
plac==1.1.3
plotly==4.3.0
pooch==1.1.1
preshed==3.0.2
prometheus-client==0.7.1
prompt-toolkit==2.0.10
protobuf==3.11.1
ptyprocess==0.6.0
pyarrow==0.12.1
pycparser==2.20
pydot==1.4.1
pyemd==0.5.1
pyglet==1.3.2
Pygments==2.5.2
pynvrtc==9.2
PyOpenGL==3.1.4
PyOpenGL-accelerate==3.1.4
pyparsing==2.4.5
Pyphen==0.9.5
pyrsistent==0.15.6
python-dateutil==2.8.1
python-slugify==4.0.0
pytz==2019.3
PyWavelets==1.1.1
PyYAML==5.2
pyzmq==18.1.1
qtconsole==4.6.0
raven==6.10.0
requests==2.22.0
requests-toolbelt==0.9.1
resampy==0.2.2
retrowrapper==0.3.0
retrying==1.3.3
scikit-image==0.16.2
scikit-learn==0.22
scikit-umfpack==0.3.2
scipy==1.3.3
seaborn==0.9.0
Send2Trash==1.5.0
six==1.13.0
sklearn==0.0
SoundFile==0.10.3.post1
spacy==2.2.3
srsly==0.2.0
sympy==1.4
tabulate==0.8.6
tensorboard==1.15.0
tensorflow==1.15.0
tensorflow-estimator==1.15.1
termcolor==1.1.0
terminado==0.8.3
testpath==0.4.4
text-unidecode==1.3
textacy==0.9.1
thinc==7.3.1
toolz==0.10.0
torch==1.6.0
torch-nightly==1.1.0.dev20190502
tornado==6.0.3
tqdm==4.40.0
traitlets==4.3.3
Unidecode==1.1.1
urllib3==1.24.3
uWSGI==2.0.18
virtualenv==16.7.8
wasabi==0.4.2
wcwidth==0.1.7
webencodings==0.5.1
Werkzeug==0.16.0
widgetsnbextension==3.5.1
wrapt==1.11.2
xgboost==0.90
zipp==0.6.0
zmq==0.0.0
